Класс вынес Невалидный сделал Вроде подразобрался, но все равно не супер уверен в знаниях (
